requirements.txt
asgiref==3.8.1
Django==5.1.5
django-multiselectfield==0.1.13
pillow==11.1.0
sqlparse==0.5.3
